requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.52
pydantic==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.27.0